move get_celery_worker_name under ansible helpers to remove import issue
