+++ commcare-cloud/commcare_cloud/commands/ansible/helpers.py
@@ -1,7 +1,10 @@
 import getpass
 import os
+
+from commcare_cloud.environment.main import get_environment
 from commcare_cloud.environment.paths import ANSIBLE_DIR
 from six.moves import shlex_quote
+from fabric.api import env
 
 DEPRECATED_ANSIBLE_ARGS = [
     '--sudo',
@@ -48,3 +51,15 @@ def get_common_ssh_args(public_vars):
     if common_ssh_args:
         cmd_parts += ('--ssh-common-args', ' '.join(shlex_quote(arg) for arg in common_ssh_args))
     return cmd_parts
+
+
+def get_celery_worker_name(comma_separated_queue_name, worker_num):
+    environment = get_environment(env.env_name)
+    environment_environment = environment.translated_app_processes_config.environment
+    project = environment.fab_settings_config.project
+    return "{project}-{environment}-celery_{comma_separated_queue_name}_{worker_num}".format(
+        project=project,
+        environment=environment_environment,
+        comma_separated_queue_name=comma_separated_queue_name,
+        worker_num=worker_num
+    )
